only Field manages data about cells

+++ Influence_Game/src/main/java/ru/nsu/fit/nsu/m_polukhin/presenter/PlayerPresenter.java
@@ -24,7 +24,7 @@ public class PlayerPresenter extends Presenter {
         view.setAttackInfo();
     }
 
-    public void setReinforceInfo(int powerRemain) {
+    public void setReinforceInfo(long powerRemain) {
         view.setReinforceInfo(powerRemain);
     }
