+++ Influence_Game/src/main/java/ru/nsu/fit/nsu/m_polukhin/presenter/StabPresenter.java
@@ -8,7 +8,7 @@ public class StabPresenter extends Presenter{
     @Override
     public void setAttackInfo() {}
     @Override
-    public void setReinforceInfo(int powerRemain) {}
+    public void setReinforceInfo(long powerRemain) {}
     @Override
     public void gameOver() {}
 }
